api_server_example.py
# ============================================
# Ejemplo de Servidor API para Beimax
# ============================================
# Este es un ejemplo de cómo configurar tu API local
# para conectar con el juego Beimax

from flask import Flask, request, jsonify
from flask_cors import CORS
import base64
import io
from PIL import Image
import numpy as np

app = Flask(__name__)
CORS(app)  # Permitir CORS para desarrollo local

# ============================================
# Aquí importarías tu modelo de IA
# ============================================
# Por ejemplo:
# from tu_modelo import IdentificadorObjetos
# modelo = IdentificadorObjetos()

# ============================================
# Endpoint de Health Check
# ============================================
@app.route('/health', methods=['GET'])
def health_check():
    """Verifica que la API esté funcionando"""
    return jsonify({
        'status': 'ok',
        'message': 'API Beimax funcionando correctamente'
    }), 200

# ============================================
# Endpoint para Identificar Objeto
# ============================================
@app.route('/api/identify', methods=['POST'])
def identify_object():
    """
    Identifica un objeto en una imagen
    
    Entrada esperada:
    {
        "image": "data:image/jpeg;base64,/9j/4AAQ...",
        "category": "kitchen"
    }
    
    Salida:
    {
        "object": "cuchillo",
        "confidence": 0.95,
        "language": "es",
        "translations": {
            "en": "knife",
            "fr": "couteau",
            "zh": "刀"
        }
    }
    """
    try:
        data = request.json
        image_data = data.get('image')
        category = data.get('category')
        
        # Decodificar imagen base64
        image = decode_base64_image(image_data)
        
        # ============================================
        # AQUÍ VA TU LÓGICA DE IA
        # ============================================
        # Ejemplo:
        # resultado = modelo.identificar(image, category)
        # object_name = resultado['objeto']
        # confidence = resultado['confianza']
        
        # Por ahora, respuesta de ejemplo
        object_name = "cuchillo"
        confidence = 0.95
        
        # Obtener traducciones
        translations = get_translations(object_name, category)
        
        return jsonify({
            'object': object_name,
            'confidence': confidence,
            'language': 'es',
            'translations': translations
        }), 200
        
    except Exception as e:
        return jsonify({
            'error': str(e),
            'message': 'Error al identificar objeto'
        }), 500

# ============================================
# Endpoint para Validar Objeto
# ============================================
@app.route('/api/validate', methods=['POST'])
def validate_object():
    """
    Valida si el objeto mostrado es el esperado
    
    Entrada esperada:
    {
        "image": "data:image/jpeg;base64,/9j/4AAQ...",
        "expected_object": "cuchillo",
        "category": "kitchen"
    }
    
    Salida:
    {
        "isCorrect": true,
        "detectedObject": "cuchillo",
        "confidence": 0.92
    }
    """
    try:
        data = request.json
        image_data = data.get('image')
        expected_object = data.get('expected_object')
        category = data.get('category')
        
        # Decodificar imagen
        image = decode_base64_image(image_data)
        
        # ============================================
        # AQUÍ VA TU LÓGICA DE IA
        # ============================================
        # Ejemplo:
        # resultado = modelo.identificar(image, category)
        # detected_object = resultado['objeto']
        # confidence = resultado['confianza']
        # is_correct = detected_object.lower() == expected_object.lower()
        
        # Por ahora, respuesta de ejemplo
        detected_object = expected_object
        confidence = 0.92
        is_correct = True
        
        return jsonify({
            'isCorrect': is_correct,
            'detectedObject': detected_object,
            'confidence': confidence
        }), 200
        
    except Exception as e:
        return jsonify({
            'error': str(e),
            'message': 'Error al validar objeto'
        }), 500

# ============================================
# Endpoint para Traducir
# ============================================
@app.route('/api/translate', methods=['POST'])
def translate_word():
    """
    Traduce una palabra de un idioma a otro
    
    Entrada esperada:
    {
        "word": "cuchillo",
        "from": "es",
        "to": "en"
    }
    
    Salida:
    {
        "translation": "knife",
        "pronunciation": "naif",
        "phonetic": "/naɪf/"
    }
    """
    try:
        data = request.json
        word = data.get('word')
        from_lang = data.get('from')
        to_lang = data.get('to')
        
        # ============================================
        # AQUÍ VA TU LÓGICA DE TRADUCCIÓN
        # ============================================
        # Puedes usar un diccionario, API de traducción, etc.
        
        translation = translate_word_helper(word, from_lang, to_lang)
        pronunciation = get_pronunciation_helper(translation, to_lang)
        phonetic = get_phonetic_helper(translation, to_lang)
        
        return jsonify({
            'translation': translation,
            'pronunciation': pronunciation,
            'phonetic': phonetic
        }), 200
        
    except Exception as e:
        return jsonify({
            'error': str(e),
            'message': 'Error al traducir palabra'
        }), 500

# ============================================
# Endpoint para Pronunciación
# ============================================
@app.route('/api/pronounce', methods=['POST'])
def get_pronunciation():
    """
    Obtiene la pronunciación de una palabra
    
    Entrada esperada:
    {
        "word": "knife",
        "language": "en"
    }
    
    Salida:
    {
        "pronunciation": "naif",
        "phonetic": "/naɪf/",
        "audioURL": null
    }
    """
    try:
        data = request.json
        word = data.get('word')
        language = data.get('language')
        
        # ============================================
        # AQUÍ VA TU LÓGICA DE PRONUNCIACIÓN
        # ============================================
        
        pronunciation = get_pronunciation_helper(word, language)
        phonetic = get_phonetic_helper(word, language)
        
        return jsonify({
            'pronunciation': pronunciation,
            'phonetic': phonetic,
            'audioURL': None  # Puedes agregar URL de audio si lo generas
        }), 200
        
    except Exception as e:
        return jsonify({
            'error': str(e),
            'message': 'Error al obtener pronunciación'
        }), 500

# ============================================
# Funciones Helper
# ============================================

def decode_base64_image(image_data):
    """Decodifica una imagen base64 a PIL Image"""
    # Remover el prefijo data:image/...;base64,
    if ',' in image_data:
        image_data = image_data.split(',')[1]
    
    # Decodificar base64
    image_bytes = base64.b64decode(image_data)
    image = Image.open(io.BytesIO(image_bytes))
    
    return image

def get_translations(word, category):
    """
    Obtiene traducciones de una palabra
    Aquí deberías implementar tu lógica de traducción
    """
    # Diccionario de ejemplo
    translations_dict = {
        'cuchillo': {
            'en': 'knife',
            'fr': 'couteau',
            'zh': '刀'
        },
        'tenedor': {
            'en': 'fork',
            'fr': 'fourchette',
            'zh': '叉'
        },
        # Agregar más traducciones...
    }
    
    return translations_dict.get(word.lower(), {
        'en': word,
        'fr': word,
        'zh': word
    })

def translate_word_helper(word, from_lang, to_lang):
    """
    Traduce una palabra
    Implementa tu lógica de traducción aquí
    """
    # Podrías usar una API como Google Translate, DeepL, etc.
    # O un modelo local de traducción
    
    # Ejemplo simple
    return word  # Retorna la palabra tal cual por ahora

def get_pronunciation_helper(word, language):
    """
    Obtiene la pronunciación de una palabra
    Implementa tu lógica aquí
    """
    # Podrías usar bibliotecas de TTS o APIs de pronunciación
    
    return word  # Simplificado

def get_phonetic_helper(word, language):
    """
    Obtiene la transcripción fonética
    Implementa tu lógica aquí
    """
    return f'/{word}/'  # Simplificado

# ============================================
# Ejecutar Servidor
# ============================================

if __name__ == '__main__':
    print("=" * 50)
    print("🤖 Servidor API Beimax iniciado")
    print("=" * 50)
    print("✅ Health check: http://localhost:5000/health")
    print("📋 Endpoints disponibles:")
    print("   - POST /api/identify")
    print("   - POST /api/validate")
    print("   - POST /api/translate")
    print("   - POST /api/pronounce")
    print("=" * 50)
    
    # Ejecutar en modo debug para desarrollo
    app.run(host='0.0.0.0', port=5000, debug=True)
